% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSCA.R
\name{RunSCA}
\alias{RunSCA}
\title{Title}
\usage{
RunSCA(
  seu_object,
  assay = NULL,
  npcs = 50L,
  n_iter = 5L,
  nbhd_size = 15L,
  model = c("wilcoxon", "binomial", "ttest"),
  metric = "euclidean",
  verbose = FALSE,
  n_tests = "auto",
  reduction.name = "sca",
  reduction.key = "SC_",
  ...
)
}
\arguments{
\item{seu_object}{Seurat object}

\item{assay}{Assay}

\item{npcs}{Desired dimensionality of the reduction.}

\item{n_iter}{Number of iterations of SCA. More iterations usually strengthens signal, stabilizing around 3-5}

\item{nbhd_size}{Size of neighborhoods used to assess the local expression of a gene. Should be smaller than the smallest subpopulation.}

\item{model}{Model used to test for local enrichment of genes, used to compute information scores. One of [“wilcoxon”,”binomial”,”ttest”], default “wilcoxon” (recommended).}

\item{metric}{Metric used to compute k-nearest neighbor graphs for SCA score computation. Default “euclidean”. See sklearn.neighbors.DistanceMetric for list of choices.}

\item{verbose}{If True, print progress.}

\item{n_tests}{Effective number of independent genes per cell, use for FWER multiple testing correction. Set to “auto” to automatically determine by bootstrapping.}

\item{reduction.name}{Name to store dimensional reduction under in the Seurat object.}

\item{reduction.key}{Dimensional reduction key, specifies the string before the number for the dimension names.}

\item{...}{Additional arguments passed to the chosen scorer.}
}
\value{
Returns a Seurat object containing an SCA representation.
}
\description{
Title
}
\examples{
data("pbmc_small")
pbmc_small <- RunSCA(pbmc_small)
DimPlot(pbmc_small, reduction = "sca", dims = 1:2)
}
